Reclassify earth lighting geometry tests as Tier 2, port calc_lighting_params (#49)
SIM_LIGHT_CIR's CSV outputs suffer from Trick scheduling lag — the logged
lighting values reflect the previous step's inputs, making them unusable
for direct numerical comparison. Reclassify the 10 geometry tests (t01-t10)
as Tier 2 in jeod_interactions, keeping only the pipeline test as Tier 3.

Port JEOD's inline calc_lighting computation into a standalone
calc_lighting_params() function in jeod_interactions::earth_lighting.
Refactor compute_earth_lighting() to call it, eliminating duplicated
occlusion/visible/lighting equations. Clamp occlusion to [0,1] to handle
floating-point edge cases at full containment.

The Tier 3 gap (no JEOD propagating sim with earth lighting) is tracked
in issue #49.
